added model kwargs

+++ types/aichat.go
@@ -5,7 +5,7 @@ type ChatLLMArgs struct {
 	PromptKeys      map[string][]byte
 	ModelType       map[string]string
 	ModelKwargs     []struct {
-		MaxTokens     int64    `json:"max_tokens"`
+		MaxTokens     int    `json:"max_tokens"`
 		Temperature   float64  `json:"temperature"`
 		Stream        bool     `json:"stream"`
 		StopSequences []string `json:"stop_sequences"`
